Make routes importable without react router
